producers/requirements.txt
fastapi
uvicorn
celery
redis
python-multipart
pydantic
python-dotenv
